+++ layouts/default.vue
@@ -56,7 +56,7 @@ export default {
   middleware: ['authenticated'],
 
   computed: {
-    ...mapState(['managementReady', 'clusterReady']),
+    ...mapState(['managementReady', 'clusterReady', 'virtualClusterReady']),
     ...mapGetters(['productId', 'clusterId', 'namespaceMode', 'isExplorer', 'currentProduct']),
     ...mapGetters({ locale: 'i18n/selectedLocaleLabel' }),
     ...mapGetters('type-map', ['activeProducts']),
@@ -191,6 +191,13 @@ export default {
       }
     },
 
+    virtualClusterReady(a, b) {
+      if ( !isEqual(a, b) ) {
+        // Immediately update because you'll see it come in later
+        this.getGroups();
+      }
+    },
+
     product(a, b) {
       if ( !isEqual(a, b) ) {
         // Immediately update because you'll see it come in later
@@ -274,13 +281,13 @@ export default {
     },
 
     getGroups() {
-      if ( !this.clusterReady ) {
+      if ( !this.clusterReady && !this.virtualClusterReady ) {
         clear(this.groups);
 
         return;
       }
 
-      const clusterId = this.$store.getters['clusterId'];
+      const clusterId = this.$store.getters['clusterId'] || this.$store.getters['virtualClusterId'];
       const currentProduct = this.$store.getters['productId'];
       const currentType = this.$route.params.resource || '';
       let namespaces = null;
@@ -486,7 +493,7 @@ export default {
 
     <div v-if="managementReady" class="dashboard-content">
       <Header />
-      <nav v-if="clusterReady" class="side-nav">
+      <nav v-if="clusterReady || virtualClusterReady" class="side-nav">
         <div class="nav">
           <template v-for="(g, idx) in groups">
             <Group
@@ -516,7 +523,7 @@ export default {
           {{ displayVersion }}
         </div>
       </nav>
-      <main v-if="clusterReady">
+      <main v-if="clusterReady || virtualClusterReady">
         <nuxt class="outlet" />
         <ActionMenu />
         <PromptRemove />
