+++ list/harvesterhci.io.dashboard/Upgrade.vue
@@ -15,9 +15,11 @@ export default {
   },
 
   async fetch() {
+    const inStore = this.$store.getters['currentProduct'].inStore;
+
     const res = await allHash({
-      upgradeVersion: this.$store.dispatch('cluster/findAll', { type: HCI.SETTING }),
-      upgrade:        this.$store.dispatch('cluster/findAll', { type: HCI.UPGRADE })
+      upgradeVersion: this.$store.dispatch(`${ inStore }/findAll`, { type: HCI.SETTING }),
+      upgrade:        this.$store.dispatch(`${ inStore }/findAll`, { type: HCI.UPGRADE })
     });
 
     this.upgrade = res.upgrade;
@@ -34,7 +36,7 @@ export default {
   },
 
   computed: {
-    ...mapGetters(['currentCluster']),
+    ...mapGetters(['currentVirtualCluster']),
 
     versionOptions() {
       const settings = this.$store.getters['cluster/all'](HCI.SETTING);
@@ -128,7 +130,7 @@ export default {
       <h1>
         <t
           k="harvester.dashboard.header"
-          :cluster="currentCluster.nameDisplay"
+          :cluster="currentVirtualCluster.nameDisplay"
         />
       </h1>
       <button v-if="versionOptions.length" type="button" class="btn bg-warning btn-sm" @click="open">
