+++ pages/index.vue
@@ -6,7 +6,7 @@ export default {
   middleware({ redirect, store } ) {
     const seenWhatsNew = store.getters['prefs/get'](SEEN_WHATS_NEW);
     const versionInfo = getVersionInfo(store);
-
+    console.log(store.getters['prefs/afterLoginRoute'], 'afterLoginRoute')
     // If this is a new version, then take the user to the home page to view the release notes
     if (versionInfo.fullVersion !== seenWhatsNew) {
       return redirect({ name: 'home' });
